Add Inline 'with' support
Overrides the Inline() method to work around an issue in Alien::Base 1.64 and later that can cause quote_spaces() in Inline::C to choke on extra compiler flags by placing those into INC rather than CCFLAGSEX.

The test will be skipped if Inline::C is unavailable. But it's added as a recommended developer prereq because it probably makes sense to have authors test this.