Set static compiler flags explicitly
For libs_static, this was being done already, and this change simply removes the duplication.

In general though, I believe the *_static flags are set automatically by Alien::Build, but I'm not sure how that works exactly and therefore am not sure if it's 100% reliable. Therefore, populating these explicitly here probably isn't a bad idea.