Add automated tests
Prereqs for configure must be listed explicitly here because Makefile.PL is currently the only source for them, and we can't run Makefile.PL without configure prereqs installed. Catch-22.

One solution would be to have a cpanfile created automatically by Makefile.PL, and to commit that cpanfile to the repository. Doable, but it'll create yet another configure dependency, on Module::CPANfile.