Add OpenSSL test
We've had occasional trouble with the linker not finding OpenSSL symbols on some smoke test systems. This test calls a function in libneo4j-omni that uses OpenSSL to verify that linking OpenSSL was successful.

Because all current functions in libneo4j-omni that use OpenSSL rely on a working Neo4j server being available, a new function neo4j_openssl_version() is added specifically for this test. It's not meant to be called by Neo4j::Client users.