Add typemap to Inline::C test
U8 is close enough to uint_fast8_t that it worked on many platforms, but at least on FreeBSD this fails with a "conflicting types" error according to CPAN Testers. This only happens for Inline::C and not for the regular XS test because Inline::C repeats the function declaration of an autowrap in the C section of the generated XS file.

To avoid the issue, we need to have the function declaration exactly identical to the one in neo4j-client.h. For that we need our own typemap because uint_fast8_t is not included in Perl's default one.

Additionally, this works around an Inline::C bug in which its default parser removed "const" from the autowrapped function return type, leading to a "discards qualifiers" warning. I think this happened because there was more than just a single modifier here ("extern const ..."). Switching parsers didn't help with this. Since "extern" is actually the default for function declarations, we can safely remove it altogether.